requirements.txt
absl-py==0.7.1
astor==0.7.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
fuzzywuzzy==0.17.0
gast==0.2.2
grpcio==1.20.0
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
malaya==2.1.0
Markdown==3.1
MarkupSafe==1.1.1
mock==2.0.0
nltk==3.4
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.2
pbr==5.1.3
protobuf==3.7.1
PySastrawi==1.2.0
PySocks==1.6.8
python-crfsuite==0.9.6
python-dateutil==2.8.0
python-dotenv==0.10.1
python-Levenshtein==0.12.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
scikit-learn==0.19.1
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
sklearn-crfsuite==0.3.6
tabulate==0.8.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
textblob==0.15.3
toolz==0.9.0
tqdm==4.31.1
tweepy==3.7.0
Unidecode==1.0.23
urllib3==1.24.1
Werkzeug==0.15.2
xgboost==0.80
